feat(sync): add sync clients (#1)
* refactor(env): create `env` file for managing configs

This commit creates an `env` file for managing environment configs for
the SDK.

* refactor(backoff): create `ExponentialBackoff` class

This commit takes the exponential backoff logic out of the
`AsyncGraphQLClient` class and moves it over to its own class. This will
be useful when implementing the sync client.

* feat(sync): add sync clients

This commit refactors a lot of the internals of the SDK so that the sync
API also works.