feat: add `query()` to sync client
This commit adds the ability to query the semantic layer using the sync
GraphQL client. The code is largely copied from the async code, but it's
very hard to generalize and would become sort of unreadable. If this
becomes a larger theme, we can think of how to refactor later.