feat: add extra fields to `SavedQuery`
Thid commit adds `Exports` and `QueryParams` to the fields of a saved
query. This makes querying for them more useful than just returning a
name, description and label.

To make this work, I had to add an extra check in the base
`GraphQLFragmentMixin` class because it didn't handle `Optional` field
well.