feat: `query` and `compile_sql` with saved queries
This commit adds `saved_query` to `query` and `compile_sql`, so it's now
possible to use them with saved queries. I added some method overrides
so that typecheckers can detect invalid usage of them, and I also
created a method which validates at runtime whether the parameters are
valid before submitting them to the servers.