test: all parameters in `query` and `compile_sql`
The tests for `query` and `compile_sql` did not test all allowed
parameters, which made the `order_by` bug go unnoticed. This commit
fixes that by adding all the remaining parameters to the tests.

Note that this commit is in a broken state since the fix hasn't been
applied yet. That will come in a future patch.