fix: `order_by` for `compile_sql` now works as expected (#49)
* test: all parameters in `query` and `compile_sql`

The tests for `query` and `compile_sql` did not test all allowed
parameters, which made the `order_by` bug go unnoticed. This commit
fixes that by adding all the remaining parameters to the tests.

Note that this commit is in a broken state since the fix hasn't been
applied yet. That will come in a future patch.

* refactor: better internal repr of query params

This commit improves our validation and representation of query
parameters, and fixes the bug with `order_by`.

We're still in an inconsistent state: we gotta propagate the changes and
use them in other classes. Will do that in the following patch.

* refactor: make protocols use the new validations

This commit makes the ADBC and GraphQL protocol implementations use the
new stricter representation for query params. It updates the tests
accordingly.

* refactor: typing public interfaces with order by

This commit updates the public `.pyi` files to ensure we use the new
order by spec

* docs: changelog

Added changelog entries related to the order by changes.