feat: allow custom granularity in all dataclasses
We introduced custom granularities, and now we need to add support for
it in the SDK. This commit is the first step towards that. It changes
all usages of `TimeGranularity` by a new `Grain` type, which can be
either the standard `TimeGranularity` or a `str` that represents a
custom granilarity.

We still haven't changed the underlying implementation to allow for
fetching these custom grains or sending them via GraphQL. This will be
next.