feat: add `extra_gql_fields` to `GraphQLFragmentMixin`
This commit adds a new classmethod called `extra_gql_fields` to
`GraphQLFragmentMixin`. This method should return any extra fields that
should be added to the GraphQL fragment that represents that dataclass.

This can be paired with `dataclasses.InitVar` to create "hidden" fields
that get queried via GraphQL and fed into `__post_init__` without
actually exposing them to the dataclass.

I used this to combine `queryableGranularities` and
`queryableTimeGranularities` in one single field that is a list of
`Grain`, so that it's less confusing to users.