fix: preload models in `__init__`
We need to call `BaseModel._register_subclasses` otherwise models will
fail to use `camelCase` and raise deprecation warnings. That is done in
`dbtsl.models.__init__`. If the user never explicitly imports that, this
won't get called, and they might get an error.

This fixes that by adding an explicit call to it on the library init.