feat: allow groupby-only queries
This commit adds support for group_by only adhoc queries with no
metrics. This has a similar effect to listing dimension values, and is
being added to the SDK for parity with what the ADBC and GraphQL APIs
can do already.

I changed the validation logic to allow for it, and I added test cases
for it.