feat: increase maximum message size
A batch can contain up to 1024 rows. If the columns contain large
amounts of data (json blogs or free text), the total batch size can
exceed the max message size. By doubling this we should be able to
accommodate most use cases.

TODO: if we hear of people reaching this max, maybe we should make it
configurable.