feat: revamp timeouts (#69)
* feat: add `timeout_ms` to GraphQL clients

This commit adds a new `timeout_ms` parameter to our sync and async
GraphQL clients, which will get passed to the underlying `gql`
transport. This allows us to set the HTTP timeout of the requests.

* docs: add changelog entry

* feat: add `TimeoutOptions` to GraphQL clients

This commit adds a new `TimeoutOptions` dataclass that can be used to
configure various timeouts for the GraphQL APIs. Users are also allowed
to pass in `int` or `float` timeouts, which will get used as the
`total_timeout`.

* docs: update changie entry

* feat: use `TimeoutOptions` for total timeout

Start using `TimeoutOptions.total_timeout` in both GraphQL clients.
We'll still fall back to `ExponentialBackoff.timeout_ms`, but that's
deprecated now.

* feat: re-raise proper errors from transport timeouts

Instead of raising `asyncio.TimeoutError` or `requests.TimeoutError`,
the SDK will raise its own `TimeoutError` if a timeout occurs in one of
the underlying GraphQL transports.

This should make it easier for end users to catch such errors without
worrying about implementation details and the underlying libraries.

* fixup! feat: use `TimeoutOptions` for total timeout

* feat: increase default total timeout to 60s

* fix: pass `execute_timeout` into the GraphQL client

* test: add integration tests for timeouts

* feat: make timeouts be upper bounded by `total_timeout`