fix: Create `FlexibleEnumMeta` enum metaclass
This commit adds a new `FlexibleEnumMeta` enum metaclass that can be
added to enums to make them accept arbitrary values as input and return
a special `UNKNOWN` entry.

I added some tests that will assert all enums exported by the SDK will
use this metaclass, and that all of them declare the `UNKNOWN` attribute
with the correct value.