refactor: make pyright and mypy strict
This commit makes both pyright and mypy operate on the strictest
setting. I had to change some code around or add type ignores in cases
where it's being too strict or when the stubs for dependencies are not
precise enough.

I only did this for the `dbtsl` package in this commit. I'll refactor
the tests in the next commit.