chore: make `pyright` run in `strict` mode and add `mypy` (#78)
* ci: add mypy

This commit introduces mypy to our pre-commit hooks and CI. This should
increase the amount of typing bugs we catch.

* refactor: make pyright and mypy strict

This commit makes both pyright and mypy operate on the strictest
setting. I had to change some code around or add type ignores in cases
where it's being too strict or when the stubs for dependencies are not
precise enough.

I only did this for the `dbtsl` package in this commit. I'll refactor
the tests in the next commit.

* test: pyright in tests

This commit changes some of our tests to work with stricter pyright
settings. I made the checks for tests a little less strict than the real
`dbtsl` tests since our tests sometimes inject wrong types into the API
to purposefully induce errors, or mocks stuff that's really hard to
type.

* docs: changelog