feat: officially support Python 3.13 (#83)
This commit adds support for Python 3.13 in the package metadata. There
were no code changes required.

As usual, we had to add some constraints to our dependencies directly.
Namely:
- `pyarrow` only supports Python 3.13 from v18.0.0 onwards
- `typing-extensions ` only supports Python 3.13 from v4.12.0 onwards

Direct dependencies are hard enforced by pip when installing the
package, and users won't be able to install if their project requires
an older version.

Additionally, I had to add constraints to two transitive dependencies:
- `frozenlist` only supports Python 3.13 from v1.5.0 onwards
- `importlib-resources` only supports Python 3.13 from v6.0.0 onwards

Transitive dependencies are NOT envorced by pip when installing the
package, and users need to manually add constraints to them. This is not
our problem: it's a problem with the dependency spec of downstream
packages. This is why our `constraints/` folder exists.

All unit/integration tests are passing.